### 2020-05-22 #2205 @whedon
[PRE REVIEW]: mmaqshiny v1.0: R-Shiny package to explore air quality mobile monitoring data
**Submitting author:** @adithirgis (<a href="http://orcid.org/0000-0002-1764-1379">Adithi Upadhya</a>)
**Repository:** <a href="https://github.com/meenakshi-kushwaha/mmaqshiny" target ="_blank">https://github.com/meenakshi-kushwaha/mmaqshiny</a>
**Version:** v.1.0.0
