### 2020-05-22 #2205 @arfon
:wave: @adithirgis - thanks for your submission to JOSS. Could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission?
