### 2020-05-22 #2205 @adithirgis
Hello Arfon,
Thank you!
Here are my suggestions:
- masalmon
- garretrc

Please let me know if I need to suggest more reviewers.

Warm regards and stay safe
Adithi R. Upadhya

On Fri, May 22, 2020 at 9:48 PM Arfon Smith <notifications@github.com>
wrote:

> 👋 @adithirgis <https://github.com/adithirgis> - thanks for your
> submission to JOSS. Could you take a look a this list of potential
> reviewers <https://bit.ly/joss-reviewers> and identify a few people who
> would be good candidates to review this submission?
>
