### 2020-05-23 #2205 @arfon
:wave: @cimentadaj, @edonnachie, @ibarraespinosa - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
