### 2020-06-01 meenakshi-kushwaha/mmaqshiny#2 @ibarraespinosa
Review for joss
I'm reviewing your paper for JOSS https://github.com/openjournals/joss-reviews/issues/2250. Really interesting work and shiny application. This will be useful for many people. However, your paper needs improvement. It looks like a manual and not like a manuscript from a scientific journal. JOSS allows to present short papers but they need to be accordingly structured.

- [x] Your [DESCRIPTION](https://raw.githubusercontent.com/meenakshi-kushwaha/mmaqshiny/master/DESCRIPTION) shows Adithi Upadhya and Meenakshi Kushwah. What's the role of Pratyush Agrawal and Sreekanth Vakacherla?
- [x] Improve the introduction and state the problem and your purpose.
- [x] Your [paper](https://github.com/openjournals/joss-papers/blob/joss.02250/joss.02250/10.21105.joss.02250.pdf) shows some paragraphs with the sentence and any paragraph should have at least three (start, development, end).
- [x] Please, include a table indicating each device, characteristics, and pollutant.
- [x] On your paper you state ' If required, user can also input linear regression coefficients for correcting the PM2.5 data.' Is this necessary/important/required? Why? Can you present an example with and without this correction?
- [x] I tried to run your application with your input data available e at https://github.com/meenakshi-kushwaha/mmaqshiny/tree/master/inst/shiny but the results were different. Why?
![2020-06-01](https://user-images.githubusercontent.com/27447280/83445025-a6a69600-a422-11ea-891e-c81db0e7e059.png)
![2020-06-01 (1)](https://user-images.githubusercontent.com/27447280/83445220-fd13d480-a422-11ea-8c7b-f06c863df377.png)
- [x] Are there any other software to do this job? Could you add a comparison?
- [x] There is API documentation and paper. According to your community guidelines should be in the API documentation [README.md](https://github.com/meenakshi-kushwaha/mmaqshiny/blob/master/README.md) and not in [paper](https://github.com/openjournals/joss-papers/blob/joss.02250/joss.02250/10.21105.joss.02250.pdf)

Cheers and stay safe
