### 2020-06-03 97fddda
added contributors, plot width change and timezone, and images for PM
