+++ man/mmaqshiny-package.Rd
@@ -35,5 +35,11 @@ Authors:
   \item Meenakshi Kushwaha \email{meenakshi@ilklabs.com}
 }
 
+Other contributors:
+\itemize{
+  \item Pratyush Agrawal \email{pratyush@ilklabs.com} [data contributor]
+  \item Sreekanth Vakacherla \email{sree.hcu@gmail.com} [author of introduction, etc., contributor]
+}
+
 }
 \keyword{internal}
