### 2020-06-04 7d2fafc
Added paper references and statement of purpose.
