+++ paper.bib
@@ -31,4 +31,13 @@
   publisher={Elsevier},
   doi={10.1016/j.atmosenv.2004.03.007}
 }
-
+@article{Salmon:2017,
+  title={rtimicropem: an R package supporting the analysis of RTI MicroPEM output files},
+  author={Salmon, Ma{\"e}lle and Vakacherla, Sreekanth and Mil{\`a}, Carles and Marshall, Julian and Tonne, Cathryn},
+  journal={Journal of Open Source Software},
+  volume={2},
+  number={16},
+  pages={333},
+  year={2017},
+  doi={: 10.21105/joss.00333}
+}
