### 2020-06-04 75fe30f
Added plots for DT
