+++ paper.md
@@ -31,7 +31,7 @@ Stationary air quality monitors do not capture spatial variations in air polluti
 
 The R-Shiny package `mmaqshiny` is for analysing, visualising and spatial-mapping of high-resolution air quality data collected by specific devices installed on a moving platform.  With the click of a button, the app generates: summary statistics, time series plots and spatial map of pollutant concentrations. This app reduces the time consumed for analysing each pollutant individually. It helps check the quality of the data at near real time (same day) and instantly visualise pollution hotspots. The time series plots of each pollutant help in understanding the temporal patterns of concentrations and performance of the instruments. This app will help both air pollution researchers and citizen scientists interested in conducting mobile measurements. 
 
-Specific details of instruments and parameters measured can be found in the “Instrument Details” section. Briefly, high frequency (1 Hz) data of  PM<sub>2.5</sub> (mass concentrations of particulate matter with size less than 2.5 microns); black carbon mass concentrations (BC); ultra-fine particle number concentrations; carbon-di-oxide; GPS coordinates and relative humidity (RH) data are collected by some popular portable instruments (TSI DustTrak-8530, Aethlabs microAeth-AE51, TSI CPC3007, LICOR Li-850, Garmin GPSMAP 64s and Omega USB RH probe, respectively), can be handled by this package. 
+Specific details of instruments and parameters measured can be found in the “Instrument Details” section. Briefly, high frequency (1 Hz) data of  PM<sub>2.5</sub> (mass concentrations of particulate matter with size less than 2.5 microns); black carbon mass concentrations (BC); ultra-fine particle number concentrations; carbon-di-oxide (CO<sub>2</sub>); GPS coordinates and relative humidity (RH) data are collected by some popular portable instruments (TSI DustTrak-8530, Aethlabs microAeth-AE51, TSI CPC3007, LICOR Li-850, Garmin GPSMAP 64s and Omega USB RH probe, respectively), can be handled by this package. 
 
 Under the hood, this app performs two basic functions - data correction and joining (location data + pollutant concentration data). As mentioned before, air-pollution measurements, especially BC from microAeth, on a mobile platform is susceptible to vibration-related noise because of the road conditions and vehicle suspension efficiency. The package incorporates an algorithm to remove this noise following @Apte:2011 method, followed by a loading correction as suggested by @Ban-Weiss:2009. PM<sub>2.5</sub> measurements from optical instruments are also sensitive to Relative Humidity (RH). To correct for RH, an algorithm is applied to DustTrak  PM<sub>2.5</sub> following @Chakrabarti:2004 method. Among other correction procedures, DustTrak PM<sub>2.5</sub> calibration equation depends on the study location. Hence, several linear calibration equations are available in literature from individual studies. Individual researchers can derive their own equation as well. Accommodating for these choices, the app allows users to input linear regression coefficients of their choice for calibrating the  PM<sub>2.5</sub> data. On-road concentrations can be much higher than ambient concentrations and sometimes even exceed the measurement range of sensors. CPC3007 has a fixed dynamic measurement range, and number concentration measurements above the range can be biased. Traditionally, this is addressed by applying a diluter (characterised by a dilution factor) to the instrument inlet. Therefore, for the ultra-fine particle number concentration data from CPC3007, provision is made for a dilution correction factor (default value is 1, implying no diluter used). After data correction, the app joins the data to generate a downloadable csv file. 
 
