+++ paper.md
@@ -46,7 +46,7 @@ The output is displayed in five different tabs.
 
 1) `Joined File` displays raw, cleaned and corrected data
 2) `Summary` displays summary statistics for each parameter
-3)  `Plots` displays interactive  time series line plots for all parameters
+3)  `Plots` displays interactive  time series line plots for raw parameters
 4) `Map` provides a spatial map for the user selected pollutant on OpenStreetMap background.
 5) `Alarm and Settings` tab displays each instruments settings and alarms (if any)
 
