### 2020-06-04 9ab6865
Made changes to table, under the hood para and last summary para
