### 2020-06-04 0ac74e7
Made changes to table, under the hood para and last summary para
