### 2020-06-04 218dd17
Updated paper.md
