### 2020-06-04 da8eb2e
Update paper.md
minor edits
