+++ paper.md
@@ -33,8 +33,7 @@ The R-Shiny package `mmaqshiny` is for analysing, visualising and spatial-mappin
 
 Specific details of instruments and parameters measured can be found in the “Instrument Details” section. Briefly, high frequency (1 Hz) data of  PM<sub>2.5</sub> (mass concentrations of particulate matter with size less than 2.5 microns); black carbon mass concentrations (BC); ultra-fine particle number concentrations; carbon-di-oxide (CO<sub>2</sub>); GPS coordinates and relative humidity (RH) data are collected by some popular portable instruments (TSI DustTrak-8530, Aethlabs microAeth-AE51, TSI CPC3007, LICOR Li-850, Garmin GPSMAP 64s and Omega USB RH probe, respectively), can be handled by this package. 
 
-<<<<<<< HEAD
-Under the hood, this app performs two basic functions - data correction and joining (location data + pollutant concentration data). As mentioned before, air-pollution measurements, especially BC from microAeth, on a mobile platform is susceptible to vibration-related noise because of the road conditions and vehicle suspension efficiency. The package incorporates an algorithm to remove this noise following @Apte:2011 method, followed by a loading correction as suggested by @Ban-Weiss:2009. PM<sub>2.5</sub> measurements from optical instruments are also sensitive to Relative Humidity (RH). To correct for RH, an algorithm is applied to DustTrak  PM<sub>2.5</sub> following @Chakrabarti:2004 method. Furthermore, for calibration of DustTrak PM<sub>2.5</sub> values, we have derived a calibration equation using a reference grade device. This  equation can vary with study location and season. Hence, several linear calibration equations are available in literature from individual studies. Researchers can derive their own equation as well. Accommodating for these choices, the app allows users to input linear regression coefficients of their choice for calibrating the  PM<sub>2.5</sub> data. Lastly, on-road concentrations can be much higher than ambient concentrations and sometimes even exceed the measurement range of sensors. CPC3007 has a fixed dynamic measurement range, and number concentration measurements beyond this range can be biased. Traditionally, this is addressed by applying a diluter (characterised by a dilution factor) to the instrument inlet. Therefore, for the ultra-fine particle number concentration data from CPC3007, provision is made for a dilution correction factor (default value is 1, implying no diluter used). After data correction, the app joins the data to generate a downloadable csv file. 
[…]
