### 2020-06-04 meenakshi-kushwaha/mmaqshiny#2 @adithirgis
Thanks for the careful review @ibarraespinosa. Overall, we have made several changes to the write up to enhance understanding. Please see our point by point responses below.

1. Your DESCRIPTION shows Adithi Upadhya and Meenakshi Kushwaha. What is the role of Pratyush Agrawal and Sreekanth Vakacherla?
_Response:
We have now added roles of Pratyush Agrawal (Data contribution) and Dr. Sreekanth Vakacherla (Co-Author, Instrument Specialist, Project Advisor) in [DESCRIPTION](https://github.com/meenakshi-kushwaha/mmaqshiny/blob/master/DESCRIPTION)._

2. Improve the introduction and state the problem and your purpose.
_Response:
We have rewritten the introduction to state the problem and the purpose._

3.  Your paper shows some paragraphs with sentences and any paragraph should have at least three (start, development, end).
_Response:
We have edited all paragraphs according to the guidelines._

4.  Please, include a table indicating each device, characteristics and pollutant.
_Response:
As suggested, we have included a table with device specifications under the “Instrument Details” section._

5. On your paper you state ' If required, user can also input linear regression coefficients for correcting the PM2.5 data.' Is this necessary/important/required? Why? Can you present an example with and without this correction?
Response:
_We have added a clarification on why regression coefficients for PM2.5 are important in the paper. An example is also added to [README.md](https://github.com/meenakshi-kushwaha/mmaqshiny/blob/master/README.md)._

6.  I treid to run your application with your input data availab e at https://github.com/meenakshi-kushwaha/mmaqshiny/tree/master/inst/shiny but the results were different. Why?
_Response:
The app requires the user to specify the data collection timezone. If not specified the app may take in the default user’s system timezone leading to different and inaccurate results. To avoid this error, we have now added a check: if the user leaves the ‘Input Timezone’ blank, this will disable the JOIN button till a valid timezone is specified. A link with a list of valid timezones is also included, default is set to Asia/Kolkata. We have also specified the plot widths now which was not specified earlier._

7. Are there any other software to do this job? Could you add a comparison?
_Response:
As of now, there are no applications which have been able to integrate the above mentioned devices. We know of another shiny app that visualizes data from a single instrument (without location data) (Salmon et al, (2017), rtimicropem: an R package supporting the analysis of RTI MicroPEM output files, Journal of Open Source Software, 2(16), 333, doi:10.21105/joss.00333). In contrast our app integrates and analyzes data from several instruments along with location data and spatial visualization._

8. There is the API documentation and the paper. According to the your community guidelines should be in the API documentation README.md and not in paper.
_Response:
We have removed the Community guidelines from [paper](https://github.com/meenakshi-kushwaha/mmaqshiny/blob/master/paper.md) and added to  [README.md](https://github.com/meenakshi-kushwaha/mmaqshiny/blob/master/README.md)._

Thanks again, Stay Safe.
