### 2020-06-05 meenakshi-kushwaha/mmaqshiny#2 @ibarraespinosa
Hello @adithirgis, I noticed your changes. Your software has been improved, as your paper as well (https://github.com/meenakshi-kushwaha/mmaqshiny/commit/da8eb2e9688d4727f8846ecfacee5d0a9c415102). Now I will present you some minor issues, which can be seen as optional.

- Your software is very useful. I liked very much the comparison between corrected and not corrected data. I am sure that if you add these figures, side by side, into your paper, it will gain much more _visibility_; obviously, including the required explanations.
- You import more than 30 packages. If you eventually submit this package to CRAN it could be rejected. In addition, managing software with many dependencies can be tough. My advice is to check the functions replace some functions with BASE in order to reduce the number of imported packages (optional).

If you further improve your paper, comment `@whedon generate pdf` here https://github.com/openjournals/joss-reviews/issues/2250

Best
