### 2020-06-06 8b157ea
Removed 6 dependencies
