+++ inst/shiny/app.R
@@ -3,7 +3,6 @@ library(Cairo)
 library(DT)
 library(plotly)
 library(zoo)
-# library(xts)
 library(data.table)
 library(stringr)
 library(dplyr)
@@ -14,17 +13,19 @@ library(keyring)
 library(leaflet)
 library(htmltools)
 library(lubridate)
-# library(rgdal)
-# library(sp)
 library(sf)
-# library(scales)
-# library(rsconnect)
 library(caTools)
 library(curl)
-# library(devtools)
 library(XML)
 library(purrr)
 library(shinyjs)
+
+# library(devtools)
+# library(xts)
+# library(scales)
+# library(rsconnect)
+# library(rgdal)
+# library(sp)
 # Sys.setenv(TMPDIR = 'C:/R/tmp_R') #Sys.getenv('TMPDIR')
 # Sys.setenv(JAVA_HOME = "C:/Program Files/Java/jre1.8.0_211/")
 # library(shinytest)
