+++ DESCRIPTION
@@ -31,29 +31,21 @@ Imports:
     Cairo,
     purrr,
     keyring,
-    scales,
     lubridate,
-    xts,
     zoo,
     curl,
     caTools,
-    rsconnect,
     ggplot2,
-    tibble,
     data.table,
     DT,
     dplyr,
     tidyr,
     readr,
-    sf,
     leaflet,
-    sp,
     stringr,
     tidyverse,
     shiny,
-    rgdal,
     XML,
-    devtools,
     usethis,
     forcats,
     shinyjs,
