+++ inst/shiny/app.R
@@ -19,9 +19,8 @@ library(curl)
 library(XML)
 library(purrr)
 library(shinyjs)
-
+library(xts)
 # library(devtools)
-# library(xts)
 # library(scales)
 # library(rsconnect)
 # library(rgdal)
