+++ DESCRIPTION
@@ -29,6 +29,7 @@ Imports:
     htmltools,
     sf,
     Cairo,
+    xts,
     purrr,
     keyring,
     lubridate,
