### 2020-06-06 a71014e
Added xts for leaflet
