+++ inst/shiny/app.R
@@ -1,6 +1,7 @@
 library(shiny)
 library(Cairo)
 library(DT)
+library(xts)
 library(plotly)
 library(zoo)
 library(data.table)
@@ -19,7 +20,7 @@ library(curl)
 library(XML)
 library(purrr)
 library(shinyjs)
-library(xts)
+
 # library(devtools)
 # library(scales)
 # library(rsconnect)
