+++ DESCRIPTION
@@ -21,7 +21,7 @@ Description: The R-Shiny package here is for analysing, visualising and spatial-
     default value is 1). The package joins the raw, cleaned and corrected data from the above said instruments and outputs as a 
     downloadable csv file. 
 Depends: 
-  R (>= 3.5.1)
+  R (>= 3.5.0)
 License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
@@ -52,7 +52,9 @@ Imports:
     shinyjs,
     plotly
 Suggests: 
-    testthat
+    testthat,
+    devtools,
+    usethis
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.1.0
 URL: https://github.com/meenakshi-kushwaha/mmaqshiny
