+++ R/mmaqshiny_run.R
@@ -1,7 +1,7 @@
 
-#' @title mmaqshiny: R-Shiny package to explore air quality mobile monitoring data
+#' @title mmaqshiny: R-Shiny Package to Explore Air-Quality Mobile Monitoring Data
 #' @description The R-Shiny package we present here is for analysing, visualising and spatial-mapping of high-resolution air quality data collected by specific devices installed on a moving platform.
-#' @keywords mmaqshiny, air quality, shiny, mobile monitoring
+#' @keywords mmaqshiny
 #' @examples
 #' \dontrun{
 #' library(mmaqshiny)
