+++ inst/shiny/app.R
@@ -1,25 +1,19 @@
 library(shiny)
 library(Cairo)
-library(xts)
 library(DT)
 library(plotly)
-library(zoo)
 library(data.table)
 library(stringr)
 library(dplyr)
 library(ggplot2)
-library(readr)
-library(tidyverse)
-library(keyring)
 library(leaflet)
-library(htmltools)
-library(lubridate)
-library(sf)
-library(caTools)
-library(curl)
 library(XML)
-library(purrr)
+library(htmltools)
 library(shinyjs)
+library(lubridate)
+library(zoo)
+library(caTools)
+library(xts)
 
 # library(devtools)
 # library(scales)
