+++ inst/shiny/app.R
@@ -20,8 +20,8 @@ library(xts)
 # library(rsconnect)
 # library(rgdal)
 # library(sp)
-# Sys.setenv(TMPDIR = 'C:/R/tmp_R') #Sys.getenv('TMPDIR')
-# Sys.setenv(JAVA_HOME = "C:/Program Files/Java/jre1.8.0_211/")
+# Sys.setenv(TMPDIR <- 'C:/R/tmp_R') #Sys.getenv('TMPDIR')
+# Sys.setenv(JAVA_HOME <- "C:/Program Files/Java/jre1.8.0_211/")
 # library(shinytest)
 # recordTest("D:/Dropbox/APMfull/Codes/GitHub Codes/ILKConsultancy")
 # deployApp()
@@ -179,7 +179,7 @@ server <- function(input, output, session) {
       return(NULL)
     }else{
       df_list <- lapply(inFile$datapath, function(y){
-        JSON_csv =read.delim(y, header=FALSE, sep=",", row.names=NULL, skip=4)
+        JSON_csv <-read.delim(y, header=FALSE, sep=",", row.names=NULL, skip=4)
         JSON_csv_date<-as.character(JSON_csv[1,2])
         JSON_csv_date<-as.Date(JSON_csv_date, format="%m/%d/%y")
         if(is.na(JSON_csv_date))
@@ -201,7 +201,7 @@ server <- function(input, output, session) {
       return(NULL)
     }
     req(input$file1)
-    files3 = lapply(inFile$datapath, function(y){
+    files3 <- lapply(inFile$datapath, function(y){
       pfile2 <-htmlTreeParse(y, error = function (...) {}, useInternalNodes = T)
       elevations <- as.numeric(xpathSApply(pfile2, path = "//trkpt/ele", xmlValue))
       times <- xpathSApply(pfile2, path = "//trkpt/time", xmlValue)
@@ -233,7 +233,7 @@ server <- function(input, output, session) {
       df_list <- lapply(inFile$datapath, function(y){
         JSON_csv_header =read.delim(y, header = FALSE, sep=",", skip = 15, row.names=NULL, stringsAsFactors = FALSE)
         JSON_csv_header<-JSON_csv_header[1,]
-        JSON_csv= read.delim(y, skip = 17, header = FALSE, sep =",")
+        JSON_csv<- read.delim(y, skip = 17, header = FALSE, sep =",")
         colnames( JSON_csv ) <- unlist(JSON_csv_header)
         JSON_csv<-completeFun(JSON_csv, c("BC"))
         JSON_csv$date1<- with(JSON_csv, as.POSIXct(paste(as.Date(Date, format="%Y/%m/%d"), Time), tz=input$timezone))
@@ -259,7 +259,7 @@ server <- function(input, output, session) {
       return(NULL)
     }else{
       df_list <- lapply(inFile$datapath, function(y){
-        JSON_csv =read.delim(y, header=TRUE, sep=",", row.names=NULL, skip=29)
+        JSON_csv <-read.delim(y, header=TRUE, sep=",", row.names=NULL, skip=29)
         JSON_csv_date<-as.character(JSON_csv[1,1])
         JSON_csv_date<-as.Date(JSON_csv_date, format="%d-%m-%Y")
         if(is.na(JSON_csv_date))
@@ -279,7 +279,7 @@ server <- function(input, output, session) {
       return(NULL)
     }else{
       df_list <- lapply(inFile$datapath, function(y){
-        JSON_csv =read.delim(y, header=TRUE, sep=";", row.names=NULL, skip=2)
[…]
