### 2020-06-17 12a799e
assignment operator and file structure
