+++ inst/shiny/app.R
@@ -15,16 +15,6 @@ library(zoo)
 library(caTools)
 library(xts)
 
-# library(devtools)
-# library(scales)
-# library(rsconnect)
-# library(rgdal)
-# library(sp)
-# Sys.setenv(TMPDIR <- 'C:/R/tmp_R') #Sys.getenv('TMPDIR')
-# Sys.setenv(JAVA_HOME <- "C:/Program Files/Java/jre1.8.0_211/")
-# library(shinytest)
-# recordTest("D:/Dropbox/APMfull/Codes/GitHub Codes/ILKConsultancy")
-# deployApp()
 
 ui <- fluidPage(
   h1("Explore Mobile Monitoring Air Pollution Data"),
@@ -170,7 +160,9 @@ ui <- fluidPage(
 
 
 server <- function(input, output, session) {
+
   options(shiny.maxRequestSize = 30*1024^2, shiny.launch.browser = TRUE)
+
   ## date verification
 
   CPC_f_date<- reactive({
@@ -600,7 +592,7 @@ server <- function(input, output, session) {
 
   data_blank<-reactive({
     if(is.null(input$file1) & is.null(input$file2) & is.null(input$file3) & is.null(input$file4) & is.null(input$file5) & is.null(input$file6)){
-      pfile2 <-htmlTreeParse("2019_09_25_h091000_KAN_Garmin_3.gpx",error = function (...) {}, useInternalNodes = T)
+      pfile2 <-htmlTreeParse("data/GPSMAP64s/2019_09_25_h091000_KAN_Garmin_3.gpx",error = function (...) {}, useInternalNodes = T)
       elevations <- as.numeric(xpathSApply(pfile2, path = "//trkpt/ele", xmlValue))
       times <- xpathSApply(pfile2, path = "//trkpt/time", xmlValue)
       coords <- xpathSApply(pfile2, path = "//trkpt", xmlAttrs)
@@ -616,7 +608,7 @@ server <- function(input, output, session) {
       setkey(GPS_f, date)
 
 
-      CO2_f<-read.delim("2019_09_25_h091000_KAN_CO2.csv", skip=1, sep=",", header=TRUE, row.names=NULL,stringsAsFactors=FALSE)
+      CO2_f<-read.delim("data/LI-COR/2019_09_25_h091000_KAN_LI_COR.csv", skip=1, sep=",", header=TRUE, row.names=NULL,stringsAsFactors=FALSE)
       CO2_f<-CO2_f[,1:3]
       names(CO2_f)<-c("Date", "Time", "CO2")
       CO2_f$date<- with(CO2_f, as.POSIXct(paste(as.Date(Date, format="%d-%m-%Y"), Time), tz="Asia/Kolkata"))#"%Y-%m-%d"; "%d-%m-%Y"
@@ -631,9 +623,9 @@ server <- function(input, output, session) {
         completeVec <- complete.cases(data[, desiredColumns])
         return(data[completeVec, ])
       }
-      BC_f_header =read.delim("2019_09_25_h091000_KAN_AE12.csv", header = FALSE, sep=",", skip = 15, row.names=NULL, stringsAsFactors = FALSE)
+      BC_f_header<-read.delim("data/AE51/2019_09_25_h091000_KAN_AE51.csv", header = FALSE, sep=",", skip = 15, row.names=NULL, stringsAsFactors = FALSE)
       BC_f_header<-BC_f_header[1,]
-      BC_f<- read.delim("2019_09_25_h091000_KAN_AE12.csv", skip = 17, header = FALSE, sep =",")
[…]
