### 2020-06-17 1dfc683
added -ve color scale for BC
