+++ inst/shiny/app.R
@@ -1424,9 +1424,8 @@ server <- function(input, output, session) {
       pal <-colorBin( "Spectral", bins=risk.bins, na.color = "#808080", reverse=TRUE)
     }
 
-    leaflet(data) %>% addProviderTiles(providers$MtbMap) %>%
-      addProviderTiles(providers$Stamen.TonerLines,
-                       options = providerTileOptions(opacity = 0.25)) %>%
+    leaflet(data) %>%
+      addProviderTiles(providers$Stamen.TonerLite) %>%
       addCircles(data=data, lng=~Longitude, lat=~Latitude, popup=  paste("Date:", data$date, "<br>",
                                                                          "AE51_BC (ug/m3):", round(as.numeric(data$BC), digits = 2), "<br>",
                                                                          "DT8530_PM2.5 (ug/m3):",round(as.numeric(data$PM2.5), digits = 2), "<br>",
