### 2020-06-17 fecbe5a
Remove - in paper and spacing
