+++ inst/shiny/app.R
@@ -19,47 +19,46 @@ library(xts)
 ui <- fluidPage(
   h1("Explore Mobile Monitoring Air Pollution Data"),
   tags$head(
-    tags$style(HTML("
-                    .sidebar { height: 10vh; overflow-y: auto; font-size: 14px;}
-                    " ,
-                    "
-                    .shiny-output-error-validation {
-                    color: red;font-size: 14px;
-                    }
-                    "
-    )) ),
+    tags$style(HTML(".sidebar {
+                    height : 10vh; overflow-y : auto; font-size : 14px;
+                    }" ,
+                    ".shiny-output-error-validation {
+                    color : red; font-size : 14px;
+                    }"
+    ))),
   sidebarLayout(position = "left",
-                sidebarPanel(width=3,
-                  conditionalPanel(condition = "input.tabs1==6",
+                sidebarPanel(width = 3,
+                  conditionalPanel(condition = "input.tabs1 == 6",
                                    tags$hr(),
                                    h4("Alarms! Check for any malfunction."),
                                    tags$hr()),
-                  conditionalPanel(condition = "input.tabs1==4",
+                  conditionalPanel(condition = "input.tabs1 == 4",
                                    tags$hr(),
                                    selectInput("palleInp", "Map this pollutant", "Select"),
                                    tags$hr()),
-                  conditionalPanel(condition = "input.tabs1==3",
+                  conditionalPanel(condition = "input.tabs1 == 3",
                                    tags$hr(),
                                    h4("Summary Statistics"),
                                    tags$hr()),
-                  conditionalPanel(condition = "input.tabs1==2",
+                  conditionalPanel(condition = "input.tabs1 == 2",
                                    tags$hr(),
                                    h4("Time series plots"),
                                    tags$hr()
                   ),
-
-                  conditionalPanel(condition = "input.tabs1==1",
+                  conditionalPanel(condition = "input.tabs1 == 1",
                                    tags$hr(),
                                    helpText("Choose mobile monitoring files."),
                                    tags$hr(),
                                    useShinyjs(),
-                                   test <- a("Input Timezone* (link to supported timezones)", href="https://en.wikipedia.org/wiki/List_of_tz_database_time_zones", style = "font-size:14px; ",target="_blank"),
-                                   textInput("timezone","", value = "Asia/Kolkata", width = NULL),
+                                   test <- a("Input Timezone* (link to supported timezones)",
+                                             href = "https://en.wikipedia.org/wiki/List_of_tz_database_time_zones",
+                                             style = "font-size:14px; ", target = "_blank"),
+                                   textInput("timezone", "", value = "Asia/Kolkata", width = NULL),
                                    tags$hr(),
                                    fileInput("file1",
                                              "GPSMAP 64s - location files",
                                              multiple = TRUE,
-                                             accept=c('.gpx')),
[…]
