+++ inst/shiny/app.R
@@ -1256,207 +1256,263 @@ server <- function(input, output, session) {
 
   output$plot5 <- renderPlotly({
     if(is.null(input$file1) & is.null(input$file2) & is.null(input$file3) & is.null(input$file4)
+       & is.null(input$file5) & is.null(input$file6)) {
+      data <- data_blank()
+      ggplotly(ggplot(data, aes(as.POSIXct(date), as.numeric(Latitude))) +
+                 geom_line(size = 0.6, color = "dodgerblue2") +
+                 labs(title = "Latitude (degree)",
+                      y = "",
+                      x = "") +
+                 scale_x_datetime( date_labels  = "%H:%M") + theme_minimal() +
+                 theme(legend.text = element_text(size = 18), plot.title = element_text(size = 14, face = "bold"),
+                       axis.title = element_text(size = 14), axis.text = element_text(size = 14, face = "bold"),
+                       panel.border = element_rect(colour = "black", fill = NA, size = 1.2))
+               )}
+    else if(is.null(GPS_f())) {}
+    else if(!is.null(GPS_f())) {
+      data <- data_joined()
+      ggplotly(ggplot(data, aes(as.POSIXct(date), as.numeric(Latitude))) +
+                 geom_line(size = 0.6, color = "dodgerblue2") +
+                 labs(title = "Latitude (degree)",
+                      y = "",
+                      x = "") +
+                 scale_x_datetime( date_labels  = "%H:%M") + theme_minimal() +
+                 theme(legend.text = element_text(size = 18), plot.title = element_text(size = 14, face = "bold"),
+                       axis.title = element_text(size = 14), axis.text = element_text(size = 14, face = "bold"),
+                       panel.border = element_rect(colour = "black", fill = NA, size = 1.2))
+               )}
+    })
+  output$plot6 <- renderPlotly({
+    if(is.null(input$file1) & is.null(input$file2) & is.null(input$file3) & is.null(input$file4)
        & is.null(input$file5) & is.null(input$file6)){
       data <- data_blank()
-      ggplotly(ggplot(data, aes(as.POSIXct(date), as.numeric(Latitude)))+ geom_line(size=0.6, color="dodgerblue2")+
-                 labs(title="Latitude (degree)",
-                      y="",
-                      x="")+scale_x_datetime( date_labels = "%H:%M")+theme_minimal()+theme(legend.text=element_text(size=18),plot.title = element_text(size = 14, face = "bold"), axis.title = element_text(size=14),axis.text = element_text(size = 14, face = "bold"),panel.border = element_rect(colour = "black", fill=NA, size=1.2))
-
-      )}
-    else if(is.null(GPS_f() )) {}
-    else if(!is.null(GPS_f() )){
+      ggplotly(ggplot(data, aes(as.POSIXct(date), as.numeric(Longitude))) +
+                 geom_line(size = 0.6, color = "dodgerblue2") +
+                 labs(title = "Longitude (degree)",
+                      y = "",
+                      x = "") +
+                 scale_x_datetime(date_labels  = "%H:%M") + theme_minimal() +
+                 theme(legend.text = element_text(size = 18), plot.title = element_text(size = 14, face = "bold"),
+                       axis.title = element_text(size = 14), axis.text = element_text(size = 14, face = "bold"),
+                       panel.border = element_rect(colour = "black", fill = NA, size = 1.2))
+               )}
+    else if(is.null(GPS_f())) {}
+    else if(!is.null(GPS_f())) {
       data <- data_joined()
[…]
