+++ inst/shiny/app.R
@@ -162,14 +162,14 @@ server <- function(input, output, session) {
 
   CPC_f_date <- reactive({
     inFile <- input$file4
-    if(is.null(input$file4)) {
+    if (is.null(input$file4)) {
       return(NULL)
     }else {
       df_list <- lapply(inFile$datapath, function(y) {
         JSON_csv <- read.delim(y, header = FALSE, sep = ",", row.names = NULL, skip = 4)
         JSON_csv_date <- as.character(JSON_csv[1, 2])
         JSON_csv_date <- as.Date(JSON_csv_date, format = "%m/%d/%y")
-        if(is.na(JSON_csv_date)) {
+        if (is.na(JSON_csv_date)) {
           JSON_csv <- read.delim(y, header = FALSE, sep = ",", row.names = NULL, skip = 4)
           JSON_csv_date <- as.character(JSON_csv[1, 2])
           JSON_csv_date <- as.Date(JSON_csv_date, format = "%m-%d-%Y")
@@ -215,11 +215,11 @@ server <- function(input, output, session) {
 
   BC_f_date <- reactive({
     inFile <- input$file2
-    if(is.null(input$file2)) {
+    if (is.null(input$file2)) {
       return(NULL)
     }else {
       completeFun <- function(data, desiredColumns) {
-        completeVec <- complete.cases(data[, desiredColumns])
+        completeVec <- complete.cases(data[ , desiredColumns])
         return(data[completeVec, ])
       }
       df_list <- lapply(inFile$datapath, function(y) {
@@ -231,7 +231,7 @@ server <- function(input, output, session) {
         JSON_csv <- completeFun(JSON_csv, c("BC"))
         JSON_csv$date1 <- with(JSON_csv, as.POSIXct(paste(as.Date(Date, format = "%Y/%m/%d"), Time),
                                                     tz = input$timezone))
-        if(is.null(JSON_csv$date1)) {
+        if (is.null(JSON_csv$date1)) {
           JSON_csv$date1 <- with(JSON_csv, as.POSIXct(paste(as.Date(Date, format = "%d-%m-%Y"), Time),
                                                       tz = input$timezone))
         }
@@ -241,7 +241,7 @@ server <- function(input, output, session) {
       BC_f <- data
       BC_f_date1 <- as.character(BC_f[1, 1])
       BC_f_date <- as.Date(BC_f_date1, format = "%d-%m-%Y")    #"%d-%m-%Y"; "%Y/%m/%d"
-      if(is.na(BC_f_date)) {
+      if (is.na(BC_f_date)) {
         BC_f_date <- as.Date(BC_f_date1, format = "%Y/%m/%d")     #"%d-%m-%Y"; "%Y/%m/%d"
       }
       return(BC_f_date)
@@ -252,14 +252,14 @@ server <- function(input, output, session) {
 
   DT_f_date <- reactive({
     inFile <- input$file3
-    if(is.null(input$file3)) {
+    if (is.null(input$file3)) {
       return(NULL)
     }else {
       df_list <- lapply(inFile$datapath, function(y) {
         JSON_csv <- read.delim(y, header = TRUE, sep = ",", row.names = NULL, skip = 29)
         JSON_csv_date <- as.character(JSON_csv[1, 1])
         JSON_csv_date <- as.Date(JSON_csv_date, format = "%d-%m-%Y")
-        if(is.na(JSON_csv_date)) {
+        if (is.na(JSON_csv_date)) {
           JSON_csv <- read.delim(y, header = TRUE, sep = ",", row.names = NULL, skip = 29)
           JSON_csv_date <- as.character(JSON_csv[1, 1])
           JSON_csv_date <- as.Date(JSON_csv_date, format = "%m/%d/%Y")
@@ -275,14 +275,14 @@ server <- function(input, output, session) {
[…]
