+++ inst/shiny/tests/test_R.R
@@ -1,27 +1,21 @@
-app <- ShinyDriver$new("../", loadTimeout = 1e+05)
+app <- ShinyDriver$new("../", loadTimeout = 10000)
 app$snapshotInit("test_R")
 
-app$setInputs(tabs1 = "3", timeout_=1000)
-app$snapshot()
-app$setInputs(tabs1 = "2", timeout_=1000)
-app$snapshot()
-app$setInputs(tabs1 = "6", timeout_=1000)
-app$snapshot()
-app$setInputs(tabs1 = "4", timeout_=1000)
-app$setInputs(tabs1 = "1", timeout_=1000)
-app$setInputs(timezone = "Asia/Kolkata")
 app$uploadFile(file1 = "2019_09_16_h094851_KAN_Garmin_3.gpx")
 app$uploadFile(file2 = "2019_09_16_h094851_KAN_AE12.csv")
 app$uploadFile(file3 = "2019_09_16_h094851_KAN_DT809.csv")
-app$setInputs(Slope = 0.21, Intercept = 11.1, timeout_=1000)
 app$uploadFile(file4 = "2019_09_16_h094851_KAN_CPC.csv")
-app$setInputs(DF = 5.5)
 app$uploadFile(file6 = "2019_09_16_h094851_KAN_CO2.csv")
-app$setInputs(join_button = "click", timeout_=1000)
-app$setInputs(tabs1 = "3", timeout_=1000)
+app$setInputs(join_button = "click", timeout_=100000)
 app$snapshot()
-app$setInputs(tabs1 = "6", timeout_=1000)
+app$setInputs(tabs1 = "3", timeout_=100000)
 app$snapshot()
-app$setInputs(tabs1 = "4", timeout_=1000)
-app$setInputs(tabs1 = "2", timeout_=1000)
+app$setInputs(tabs1 = "6", timeout_=100000)
+app$setInputs(tabs1 = "2", timeout_=100000)
+app$snapshot()
+app$setInputs(tabs1 = "4", timeout_=100000)
+app$setInputs(tabs1 = "6", timeout_=100000)
+app$snapshot()
+app$setInputs(tabs1 = "4", timeout_=100000)
+app$snapshot()
 app$snapshot()
