+++ inst/shiny/app.R
@@ -158,7 +158,7 @@ server <- function(input, output, session) {
 
   options(shiny.maxRequestSize = 30*1024^2, shiny.launch.browser = TRUE)
 
-  ## CPC date
+  ## date matching
 
   CPC_f_date <- reactive({
     inFile <- input$file4
@@ -180,8 +180,6 @@ server <- function(input, output, session) {
     CPC_f2_date <- df_list[[1]]
     return(CPC_f2_date)
   })
-
-  ## GPS date
 
   GPS_f_date <- reactive({
     trial  <- data.frame()
@@ -210,8 +208,6 @@ server <- function(input, output, session) {
     GPS_f_date <- as.Date(GPS_f_date, format="%Y-%m-%d") #"%m/%d/%Y"
     return(GPS_f_date)
   })
-
-  ## AE51 date
 
   BC_f_date <- reactive({
     inFile <- input$file2
@@ -248,8 +244,6 @@ server <- function(input, output, session) {
     }
   })
 
-  ## DT8530 date
-
   DT_f_date <- reactive({
     inFile <- input$file3
     if (is.null(input$file3)) {
@@ -271,8 +265,6 @@ server <- function(input, output, session) {
     }
   })
 
-  ## LI-COR date
-
   CO2_f_date <- reactive({
     inFile <- input$file6
     if (is.null(input$file6)) {
@@ -293,6 +285,7 @@ server <- function(input, output, session) {
       return(CO2_f_date)
     }
     })
+
 
   ## GPS and pollutant file joining
 
@@ -491,6 +484,7 @@ server <- function(input, output, session) {
 
   })
 
+
   ## file name matching
 
   file_name_CPC <- reactive({
@@ -570,6 +564,7 @@ server <- function(input, output, session) {
       return(name_CO2)
     }
   })
+
   file_name_RH <- reactive({
     inFile <- input$file5
     if (is.null(inFile)) {
@@ -580,6 +575,7 @@ server <- function(input, output, session) {
       return(name_RH)
     }
   })
+
   file_name_GPS <- reactive({
     inFile <- input$file1
     if (is.null(inFile)) {
@@ -590,6 +586,7 @@ server <- function(input, output, session) {
       return(name_GPS)
     }
   })
+
   file_name_BC <- reactive({
     inFile <- input$file2
     if (is.null(inFile)) {
@@ -600,6 +597,7 @@ server <- function(input, output, session) {
       return(name_BC)
     }
   })
+
   file_name_DT <- reactive({
     inFile <- input$file3
     if (is.null(inFile)) {
@@ -611,7 +609,8 @@ server <- function(input, output, session) {
     }
   })
 
- ## preloaded table
+
+  ## preloaded table
 
   data_blank <- reactive({
     if (is.null(input$file1) & is.null(input$file2) & is.null(input$file3) &
@@ -866,6 +865,7 @@ server <- function(input, output, session) {
     setkey(joined_GPS_CO2, date)
     return(joined_GPS_CO2)
   })
+
   joined_GPS_DT <- reactive({
     name_GPS <- file_name_GPS()
     name_DT  <- file_name_DT()
@@ -896,6 +896,7 @@ server <- function(input, output, session) {
     setkey(joined_GPS_DT, date)
     return(joined_GPS_DT)
   })
+
   DT_RH <- reactive({
     name_RH <- file_name_RH()
     name_DT <- file_name_DT()
@@ -941,6 +942,7 @@ server <- function(input, output, session) {
     setkey(DT_f, date)
     return(DT_f)
   })
+
   joined_GPS_BC <- reactive({
     name_GPS <- file_name_GPS()
     name_BC <- file_name_BC()
[…]
