+++ inst/shiny/tests/test_m.R
@@ -1,0 +1,15 @@
+app <- ShinyDriver$new("../", loadTimeout = 1e+05)
+app$snapshotInit("test_m")
+
+app$uploadFile(file1 = "2019_09_16_h094851_KAN_Garmin_3.gpx")
+app$uploadFile(file3 = "2019_09_16_h094851_KAN_DT809.csv")
+app$setInputs(Slope = 0.21, Intercept = 11.1)
+app$uploadFile(file4 = "2019_09_16_h094851_KAN_CPC.csv")
+app$setInputs(join_button = "click")
+app$snapshot()
+app$setInputs(tabs1 = "3")
+app$snapshot()
+app$setInputs(tabs1 = "1")
+app$setInputs(timezone = "UTC")
+app$setInputs(join_button = "click")
+app$snapshot()
