### 2020-06-18 4ac2fd1
Added punctuation change fix 4 in Issue #3
