+++ inst/shiny/app.R
@@ -34,7 +34,8 @@ ui <- fluidPage(
                                    tags$hr()),
                   conditionalPanel(condition = "input.tabs1 == 4",
                                    tags$hr(),
-                                   selectInput("palleInp", "Map this pollutant", "Select"),
+                                   selectInput("palleInp", "Map this pollutant",
+                                               "Select"),
                                    tags$hr()),
                   conditionalPanel(condition = "input.tabs1 == 3",
                                    tags$hr(),
@@ -50,10 +51,13 @@ ui <- fluidPage(
                                    helpText("Choose mobile monitoring files."),
                                    tags$hr(),
                                    useShinyjs(),
-                                   test <- a("Input Timezone* (link to supported timezones)",
+                                   test <- a("Input Timezone* (link to supported
+                                             timezones)",
                                              href = "https://en.wikipedia.org/wiki/List_of_tz_database_time_zones",
-                                             style = "font-size:14px; ", target = "_blank"),
-                                   textInput("timezone", "", value = "Asia/Kolkata", width = NULL),
+                                             style = "font-size:14px; ",
+                                             target = "_blank"),
+                                   textInput("timezone", "", value = "Asia/Kolkata",
+                                             width = NULL),
                                    tags$hr(),
                                    fileInput("file1",
                                              "GPSMAP 64s - location files",
@@ -63,13 +67,15 @@ ui <- fluidPage(
                                    fileInput("file2", "AE51 - BC files",
                                              multiple = TRUE,
                                              accept = c("text/csv",
-                                                        "text/comma-separated-values,text/plain",
+                                                        "text/comma-separated-values,
+                                                        text/plain",
                                                         ".csv")),
                                    tags$hr(),
                                    fileInput("file3", "DT8530 - PM2.5 files",
                                              multiple = TRUE,
                                              accept = c("text/csv",
-                                                        "text/comma-separated-values,text/plain",
+                                                        "text/comma-separated-values,
+                                                        text/plain",
                                                         ".csv")),
                                    h5("Reference Correction for DT8530 - PM2.5"),
                                    numericInput("Slope",
@@ -82,24 +88,29 @@ ui <- fluidPage(
                                    fileInput("file5", "RH files",
                                              multiple = TRUE,
                                              accept = c("text/csv",
-                                                        "text/comma-separated-values,text/plain",
+                                                        "text/comma-separated-values,
+                                                        text/plain",
                                                         ".csv")),
 
                                    tags$hr(),
-                                   fileInput("file4", "CPC3007 -Particle Concentration files",
[…]
