+++ paper.md
@@ -29,11 +29,11 @@ affiliations:
 
 Stationary air quality monitors do not capture spatial variations in air pollution. Mobile monitoring or “sensors on a mobile platform” is an increasingly popular approach to measure high resolution pollution data at the street level. Coupled with location data, spatial visualization of air quality parameters helps locate localized areas of high air pollution, also called hotspots. In this approach, portable sensors are mounted on a vehicle and driven on predetermined routes to collect high frequency data (1 Hz). Analyzing this data involves cleaning and combining location data with pollutant data from different instruments. Some instruments are sensitive to factors like signal attenuation, humidity, vibrations and require additional correction algorithms unique to each instrument. A typical measurement campaign involves collecting millions of data points and a high burden of quality assurance and quality control. Our package attempts to automate and simplify the process using a shiny app.
 
-The R-Shiny package `mmaqshiny` is for analysing, visualising and spatial-mapping of high-resolution air quality data collected by specific devices installed on a moving platform.  With the click of a button, the app generates: summary statistics, time series plots and spatial map of pollutant concentrations. This app reduces the time consumed for analysing each pollutant individually. It helps check the quality of the data at near real time (same day) and instantly visualise pollution hotspots. The time series plots of each pollutant help in understanding the temporal patterns of concentrations and performance of the instruments. This app will help both air pollution researchers and citizen scientists interested in conducting mobile measurements. 
+The R-Shiny package `mmaqshiny` is for analysing, visualising and spatial mapping of high-resolution air quality data collected by specific devices installed on a moving platform.  With the click of a button, the app generates: summary statistics, time series plots and spatial map of pollutant concentrations. This app reduces the time consumed for analysing each pollutant individually. It helps check the quality of the data at near real time (same day) and instantly visualise pollution hotspots. The time series plots of each pollutant help in understanding the temporal patterns of concentrations and performance of the instruments. This app will help both air pollution researchers and citizen scientists interested in conducting mobile measurements. 
 
-Specific details of instruments and parameters measured can be found in the “Instrument Details” section. Briefly, high frequency (1 Hz) data of  PM<sub>2.5</sub> (mass concentrations of particulate matter with size less than 2.5 microns); black carbon mass concentrations (BC); ultra-fine particle number concentrations; carbon-di-oxide (CO<sub>2</sub>); GPS coordinates and relative humidity (RH) data are collected by some popular portable instruments (TSI DustTrak-8530, Aethlabs microAeth-AE51, TSI CPC3007, LICOR Li-850, Garmin GPSMAP 64s and Omega USB RH probe, respectively), can be handled by this package. 
+Specific details of instruments and parameters measured can be found in the “Instrument Details” section. `mmaqshiny` package can handle high frequency (1 Hz) measurements collected by some popular portable instruments. Specifically, the following parameters are measured : PM2.5 (mass concentrations of particulate matter with size less than 2.5 microns), black carbon mass concentrations (BC), ultra-fine particle number concentrations; carbon dioxide (CO<sub>2</sub>), GPS coordinates and relative humidity (RH).
 
[…]
