### 2020-06-18 a0c07f7
final app  changes
