### 2020-06-18 e7950d5
Update paper.md
