### 2020-06-18 b5f683e
removed two licenses
