+++ man/mmaqshiny_run.Rd
@@ -2,12 +2,14 @@
 % Please edit documentation in R/mmaqshiny_run.R
 \name{mmaqshiny_run}
 \alias{mmaqshiny_run}
-\title{mmaqshiny: R-Shiny Package to Explore Air-Quality Mobile Monitoring Data}
+\title{mmaqshiny: Explore Air Quality Mobile Monitoring Data}
 \usage{
 mmaqshiny_run()
 }
 \description{
-The R-Shiny package we present here is for analysing, visualising and spatial-mapping of high-resolution air quality data collected by specific devices installed on a moving platform.
+The package is for analysing, visualising and spatial mapping of
+high-resolution air quality data collected by specific devices installed on
+a moving platform.
 }
 \examples{
 \dontrun{
