+++ inst/shiny/app.R
@@ -91,7 +91,6 @@ ui <- fluidPage(
                                                         "text/comma-separated-values,
                                                         text/plain",
                                                         ".csv")),
-
                                    tags$hr(),
                                    fileInput("file4", "CPC3007 -
                                              Particle Concentration files",
