+++ inst/shiny/app.R
@@ -215,7 +215,7 @@ server <- function(input, output, session) {
     trial <- do.call(rbind, files3)
     GPS_f <- trial
     GPS_f <- mutate(GPS_f, date = with_tz(ymd_hms(time), input$timezone))
-    GPS_f <- GPS_f[!duplicated(as.POSIXct(GPS_f$date)),]
+    GPS_f <- GPS_f[!duplicated(as.POSIXct(GPS_f$date)), ]
     GPS_f <- GPS_f %>%
       dplyr::select(date, latitude, longitude)
     names(GPS_f) <- c("date", "latitude", "longitude")
@@ -237,7 +237,7 @@ server <- function(input, output, session) {
         JSON_csv_header <- read.delim(y, header = FALSE, sep = ",",
                                       skip = 15, row.names = NULL,
                                       stringsAsFactors = FALSE)
-        JSON_csv_header <-JSON_csv_header[1, ]
+        JSON_csv_header <- JSON_csv_header[1, ]
         JSON_csv <- read.delim(y, skip = 17, header = FALSE, sep = ",")
         colnames(JSON_csv) <- unlist(JSON_csv_header)
         JSON_csv <- completeFun(JSON_csv, c("BC"))
@@ -293,7 +293,7 @@ server <- function(input, output, session) {
       return(NULL)
     }else {
       df_list <- lapply(inFile$datapath, function(y) {
-        JSON_csv <-read.delim(y, header = TRUE, sep = ";",
+        JSON_csv <- read.delim(y, header = TRUE, sep = ";",
                               row.names = NULL, skip = 2)
         JSON_csv_date <- as.character(JSON_csv[1, 1])
         JSON_csv_date <- as.Date(JSON_csv_date, format = "%d-%m-%Y")
@@ -331,7 +331,7 @@ server <- function(input, output, session) {
       CO2_f$date <- with(CO2_f,
                          as.POSIXct(paste(
                            as.Date(Date, format = "%d-%m-%Y"), Time),
-                                           tz = input$timezone))#"%Y-%m-%d""%d-%m-%Y"
+                                           tz = input$timezone)) #"%Y-%m-%d""%d-%m-%Y"
       CO2_f <- CO2_f %>%
         dplyr::select(date, CO2)
       names(CO2_f) <- c("date", "CO2")
@@ -363,7 +363,7 @@ server <- function(input, output, session) {
     trial <- do.call(rbind, files3)
     GPS_f <- trial
     GPS_f <- mutate(GPS_f, date = with_tz(ymd_hms(time), input$timezone))
-    GPS_f <- GPS_f[!duplicated(as.POSIXct(GPS_f$date)),]
+    GPS_f <- GPS_f[!duplicated(as.POSIXct(GPS_f$date)), ]
     GPS_f <- GPS_f %>%
       dplyr::select(date, latitude, longitude)
     names(GPS_f) <- c("date", "latitude", "longitude")
@@ -387,7 +387,7 @@ server <- function(input, output, session) {
                                       stringsAsFactors = FALSE)
         JSON_csv_header <- JSON_csv_header[1, ]
         JSON_csv <- read.delim(y, skip = 17, header = FALSE, sep = ",")
-        colnames( JSON_csv ) <- unlist(JSON_csv_header)
+        colnames(JSON_csv) <- unlist(JSON_csv_header)
         JSON_csv <- completeFun(JSON_csv, c("BC"))
         JSON_csv$date1 <- with(JSON_csv,
[…]
