+++ R/mmaqshiny_run.R
@@ -1,6 +1,6 @@
 
 #' @title mmaqshiny: Explore Air Quality Mobile Monitoring Data
-#' @description The package is for analysing, visualising and spatial mapping of
+#' @description The package is for analysing, visualising and spatial plotting of
 #' high-resolution air quality data collected by specific devices installed on
 #' a moving platform.
 #' @keywords mmaqshiny
