### 2020-06-18 6e8f810
Cleaner code
