+++ inst/shiny/app.R
@@ -413,18 +413,18 @@ server <- function(input, output, session) {
       names(ef_file) <- c("Date", "ATN", "BC")
       ef_file$BC1 <- (ef_file$BC / 1000)
       BC_Final <- ef_file
-      ef_file$LD   <- ef_file$BC1 - rollapply(ef_file$BC1 , FUN = mean,
+      ef_file$LD <- ef_file$BC1 - rollapply(ef_file$BC1, FUN = mean,
                                               width = 30, align = "center",
                                               partial = TRUE)
       ef_file$LD25 <- runquantile(ef_file$LD, 300, 0.25, type = 2,
                                   endrule = c("NA"))
       ef_file$LD75 <- runquantile(ef_file$LD, 300, 0.75, type = 2,
                                   endrule = c("NA"))
-      ef_file$BC2  <- ef_file$BC1
-      ef_file$BC3  <- ef_file$BC1
+      ef_file$BC2 <- ef_file$BC1
+      ef_file$BC3 <- ef_file$BC1
       ef_file$BC2[ef_file$BC2 >= 0] <- 0
       ef_file$BC2[ef_file$BC2 < 0] <- 1
-      ef_file$BC2 <- rollapply(ef_file$BC2 , FUN = mean, width = 5,
+      ef_file$BC2 <- rollapply(ef_file$BC2, FUN = mean, width = 5,
                                align = "center",  partial = TRUE)
       Lower_LD25  <- ((ef_file$LD) > 5 * (ef_file$LD75))
       Higher_LD75 <- ((ef_file$LD) < 5 * (ef_file$LD25))
@@ -432,7 +432,7 @@ server <- function(input, output, session) {
       CEV <- data.frame(ef_file$Date, ef_file$cev1)
       CEV$ef_file.cev1[!is.na(CEV$ef_file.cev1)] <- 1
       CEV <- data.frame(CEV)
-      date_file <- data.frame(ef_file$Date,  ef_file$BC2, ef_file$BC3)
+      date_file <- data.frame(ef_file$Date, ef_file$BC2, ef_file$BC3)
       completeFun <- function(data, desiredColumns) {
         completeVec <- complete.cases(data[ , desiredColumns])
         return(data[completeVec, ])
@@ -470,7 +470,7 @@ server <- function(input, output, session) {
       }
       BC$BC_Fi <- BC$BC_Factor * BC$BC1
       BC$BC_Fi[BC$BC_Fi == 0] <- NA
-      BC$Tr <- exp(- BC$ATN /  100)
+      BC$Tr <- exp(- BC$ATN / 100)
       BC$CF <- 1 / (0.88 * BC$Tr + 0.12)
       BC$BC_Final <- BC$BC_Fi * BC$CF
       BC$BC_Fi[BC$BC_Fi < 0] <- NA
@@ -696,11 +696,11 @@ server <- function(input, output, session) {
                                 sep = ",", skip = 15, row.names = NULL,
                                 stringsAsFactors = FALSE)
       BC_f_header <- BC_f_header[1, ]
-      BC_f<- read.delim("data/AE51/2019_09_25_h091000_KAN_AE51.csv", skip = 17,
+      BC_f <- read.delim("data/AE51/2019_09_25_h091000_KAN_AE51.csv", skip = 17,
                         header = FALSE, sep = ",")
       colnames(BC_f) <- unlist(BC_f_header)
-      BC_f<-completeFun(BC_f, c("BC"))
-      BC_f$date1<- with(BC_f,
+      BC_f <-completeFun(BC_f, c("BC"))
+      BC_f$date1 <- with(BC_f,
                         as.POSIXct(paste(as.Date(Date, format="%Y/%m/%d"),
[…]
