+++ DESCRIPTION
@@ -49,7 +49,8 @@ Imports:
 Suggests: 
     testthat,
     devtools,
-    usethis
+    usethis,
+    shinytest
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.1.0
 URL: https://github.com/meenakshi-kushwaha/mmaqshiny
