### 2020-06-19 975a129
added test_map_plot
