+++ inst/shiny/tests/test_map_plot.R
@@ -1,0 +1,19 @@
+app <- ShinyDriver$new("../", seed = 9, loadTimeout = 10000)
+app$snapshotInit("test_map_plot")
+
+app$uploadFile(file1 = "2019_09_16_h094851_KAN_Garmin_3.gpx")
+app$uploadFile(file2 = "2019_09_16_h094851_KAN_AE12.csv")
+app$uploadFile(file3 = "2019_09_16_h094851_KAN_DT809.csv")
+app$setInputs(Slope = 0.21, Intercept = 11.1, wait_ = FALSE, values_ = FALSE)
+app$setInputs(join_button = "click", timeout_ = 100000)
+app$setInputs(tabs1 = "3")
+Sys.sleep(10)
+app$snapshot()
+app$setInputs(tabs1 = "2")
+Sys.sleep(10)
+app$snapshot()
+app$setInputs(tabs1 = "4")
+Sys.sleep(10)
+app$snapshot()
+app$setInputs(tabs1 = "6")
+app$snapshot()
