+++ inst/shiny/tests/test_map_plot.R
@@ -13,7 +13,7 @@ app$setInputs(tabs1 = "2")
 Sys.sleep(10)
 app$snapshot()
 app$setInputs(tabs1 = "4")
-Sys.sleep(10)
+Sys.sleep(11)
 app$snapshot()
 app$setInputs(tabs1 = "6")
 app$snapshot()
