+++ inst/shiny/tests/test_win_time.R
@@ -10,6 +10,7 @@ app$setInputs(join_button = "click", timeout_ = 100000)
 Sys.sleep(10)
 app$snapshot()
 app$setInputs(tabs1 = "3")
+Sys.sleep(10)
 app$snapshot()
 app$setInputs(tabs1 = "1")
 app$setInputs(timezone = "Asia/Kolkata")
