### 2020-06-21 a084819
added map test in win
