+++ inst/shiny/app.R
@@ -55,8 +55,10 @@ ui <- fluidPage(
                                              href = "https://en.wikipedia.org/wiki/List_of_tz_database_time_zones",
                                              style = "font-size:14px; ",
                                              target = "_blank"),
-                                   textInput("timezone", "", value = "Asia/Kolkata",
-                                             width = NULL),
+                                   selectInput("timezone",
+                                               label = "",
+                                               choices = OlsonNames(),
+                                               selected = "Asia/Kolkata"),
                                    tags$hr(),
                                    fileInput("file1",
                                              "GPSMAP 64s - location files",
