+++ inst/shiny/tests/test_mac_join.R
@@ -1,0 +1,19 @@
+app <- ShinyDriver$new("../", seed = 4, loadTimeout = 1e+05)
+app$snapshotInit("test_mac_join")
+
+Sys.sleep(10)
+app$snapshot()
+app$setInputs(tabs1 = "3")
+Sys.sleep(10)
+app$snapshot()
+app$setInputs(tabs1 = "1")
+app$uploadFile(file1 = "2019_09_16_h094851_KAN_Garmin_3.gpx")
+app$uploadFile(file2 = "2019_09_16_h094851_KAN_AE12.csv")
+app$uploadFile(file3 = "2019_09_16_h094851_KAN_DT809.csv")
+app$setInputs(Slope = 0.21, Intercept = 11.1, wait_ = FALSE, values_ = FALSE)
+app$setInputs(join_button = "click", timeout_ = 100000)
+Sys.sleep(10)
+app$snapshot()
+app$setInputs(tabs1 = "3")
+Sys.sleep(7)
+app$snapshot()
