+++ inst/shiny/tests/test_mac.R
@@ -1,41 +1,0 @@
-app <- ShinyDriver$new("../", loadTimeout = 100000)
-app$snapshotInit("test_mac")
-
-app$setInputs(timezone = "Asa")
-app$setInputs(timezone = "As")
-app$setInputs(timezone = "Asia")
-app$setInputs(timezone = "Asia/")
-app$setInputs(timezone = "Asia/K")
-app$setInputs(timezone = "Asia/Ko")
-app$setInputs(timezone = "Asia/Kol")
-app$setInputs(timezone = "Asia/Kolka")
-app$setInputs(timezone = "Asia/Kolkata")
-app$uploadFile(file1 = "2019_09_16_h094851_KAN_Garmin_3.gpx")
-app$uploadFile(file2 = "2019_09_16_h094851_KAN_AE12.csv")
-app$uploadFile(file3 = "2019_09_16_h094851_KAN_DT809.csv")
-app$setInputs(Slope = 0)
-app$setInputs(Slope = 0.3)
-app$setInputs(Slope = 0.33)
-app$setInputs(Intercept = 11)
-app$setInputs(Intercept = 111)
-app$setInputs(Intercept = 11)
-app$setInputs(Intercept = 11.1)
-app$setInputs(Slope = 0.21)
-app$uploadFile(file4 = "2019_09_16_h094851_KAN_CPC.csv")
-app$setInputs(DF = 5)
-app$setInputs(DF = 5.5)
-app$setInputs(join_button = "click", timeout_=100000)
-app$snapshot()
-app$setInputs(tabs1 = "3", timeout_=100000)
-app$snapshot()
-app$setInputs(tabs1 = "2", timeout_=100000)
-app$snapshot()
-app$setInputs(tabs1 = "1", timeout_=100000)
-app$setInputs(tabs1 = "4", timeout_=100000)
-app$snapshot()
-app$setInputs(palleInp = "PM2.5", timeout_=100000)
-app$snapshot()
-app$setInputs(palleInp = "Particle_conc", timeout_=100000)
-app$snapshot()
-app$setInputs(tabs1 = "6", timeout_=100000)
-app$snapshot()
