+++ R/mmaqshiny_run.R
@@ -6,7 +6,7 @@
 #' helps detect localized areas of high air-pollution, also called hotspots.
 #' In this approach, portable sensors are mounted on a vehicle and driven on
 #' predetermined routes to collect high frequency data (1 Hz). The package is
-#' for analysing, visualising and spatial plotting of high-resolution air-quality
+#' for analysing, visualising and spatial maps of high-resolution air-quality
 #' data collected by specific devices installed on a moving platform.
 #' @keywords mmaqshiny
 #' @examples
