+++ man/mmaqshiny-package.Rd
@@ -12,7 +12,7 @@ Mobile-monitoring or “sensors on a mobile platform”, is an increasingly
     helps detect localized areas of high air-pollution, also called hotspots. 
     In this approach, portable sensors are mounted on a vehicle and driven on 
     predetermined routes to collect high frequency data (1 Hz). 
-    mmaqshiny is for analysing, visualising and spatial plotting of 
+    mmaqshiny is for analysing, visualising and spatial mapping of 
     high-resolution air-quality data collected by specific devices installed on 
     a moving platform. 1 Hz data of PM2.5 (mass concentrations of particulate  
     matter with size less than 2.5 microns), Black carbon mass concentrations 
@@ -40,11 +40,11 @@ Useful links:
 
 }
 \author{
-\strong{Maintainer}: Adithi Upadhya \email{adithi@ilklabs.com}
+\strong{Maintainer}: Adithi R. Upadhya \email{adithi@ilklabs.com} [copyright holder]
 
 Authors:
 \itemize{
-  \item Meenakshi Kushwaha \email{meenakshi@ilklabs.com}
+  \item Meenakshi Kushwaha \email{meenakshi@ilklabs.com} [copyright holder]
 }
 
 Other contributors:
