+++ DESCRIPTION
@@ -2,17 +2,17 @@ Package: mmaqshiny
 Title: Explore Air-Quality Mobile-Monitoring Data
 Version: 1.0.0
 Authors@R:  c(
-    person("Adithi", "Upadhya", email = "adithi@ilklabs.com", role = c("aut", "cre")),
-    person("Pratyush", "Agrawal", email = "pratyush@ilklabs.com", role = "dtc"),
-    person("Sreekanth", "Vakacherla", email = "sree.hcu@gmail.com", role = c("aui", "ctb")),
-    person("Meenakshi", "Kushwaha", email = "meenakshi@ilklabs.com",  role = "aut"))
+    person(given = "Adithi R.", family = "Upadhya", email = "adithi@ilklabs.com", role = c("aut", "cre", "cph")),
+    person(given = "Pratyush", family = "Agrawal", email = "pratyush@ilklabs.com", role = "dtc"),
+    person(given = "Sreekanth", family = "Vakacherla", email = "sree.hcu@gmail.com", role = c("aui", "ctb")),
+    person(given = "Meenakshi", family = "Kushwaha", email = "meenakshi@ilklabs.com", role = c("aut", "cph")))
 Description: Mobile-monitoring or “sensors on a mobile platform”, is an increasingly 
     popular approach to measure high-resolution pollution data at the street level. 
     Coupled with location data, spatial visualisation of air-quality parameters 
     helps detect localized areas of high air-pollution, also called hotspots. 
     In this approach, portable sensors are mounted on a vehicle and driven on 
     predetermined routes to collect high frequency data (1 Hz). 
-    mmaqshiny is for analysing, visualising and spatial plotting of 
+    mmaqshiny is for analysing, visualising and spatial mapping of 
     high-resolution air-quality data collected by specific devices installed on 
     a moving platform. 1 Hz data of PM2.5 (mass concentrations of particulate  
     matter with size less than 2.5 microns), Black carbon mass concentrations 
