### 2020-06-25 #2250 @nistara
Hello @arfon, I've completed my review as well. It's a helpful software which also exposed me to the world of air pollution data!
Thank you and best wishes,
Nistara
