+++ paper.md
@@ -36,7 +36,7 @@ Specific details of instruments and parameters measured can be found in the “I
 
 Under the hood, this app performs two basic functions: data correction and joining (location data + pollutant concentration data). As mentioned earlier, air-pollution measurements on a mobile platform, especially BC from microAeth are susceptible to vibration-related noise due to the road conditions and vehicle suspension efficiency. The package incorporates an algorithm to remove this noise employing the method recommended by @Apte:2011, followed by a loading correction as suggested by @Ban-Weiss:2009. Since  PM₂.₅ measurements from optical instruments are also sensitive to Relative Humidity (RH), an algorithm is applied to DustTrak PM₂.₅ to correct for RH, following the method suggested by  @Chakrabarti:2004. Furthermore, for calibration of DustTrak PM₂.₅ values, we have derived a calibration equation using a reference grade device. This  equation can vary with study location and season. Hence, several linear calibration equations from individual studies can be available in literature. Researchers can derive their own equation as well. Accommodating for these options, the app allows users to input linear regression coefficients of their choice for calibrating the  PM₂.₅ data. Often, on-road concentrations can be much higher than ambient concentrations and sometimes even exceed the measurement range of sensors. CPC3007 has a fixed dynamic measurement range, and number concentration measurements beyond this range can be biased. Traditionally, this is addressed by adding a diluter (characterised by a dilution factor) to the instrument inlet. Therefore, for the ultra-fine particle number concentration data from CPC3007, provision is made for a dilution correction factor (default value is 1, implying no diluter used). After data correction, the app joins the data to generate a downloadable csv file. 
 
-The app is designed to accept multiple files for each parameter from a single day of measurements. The input file names should have a date prefix of the format `yyyy_mm_dd`. The app uses this prefix to check that all input files are from the same date and generates an error message if the prefixes don’t match. The package requires GPS file (.gpx) as a mandatory input along with timezone which can be selected using a dropdown menu(a link to all accepted timezone formats in R is included). All other pollutant files are optional.
+The app is designed to accept multiple files for each parameter from a single day of measurements. The input file names should have a date prefix of the format `yyyy_mm_dd`. The app uses this prefix to check that all input files are from the same date and generates an error message if the prefixes don’t match. The package requires GPS file (.gpx) as a mandatory input along with timezone which can be selected using a dropdown menu (a link to all accepted timezone in R is included). All other pollutant files are optional.
 
 Mobile-monitoring is a relatively new technique and the instruments used are also highly customised. To the authors’ knowledge, currently there is no application that can integrate the above mentioned devices, though there are a few apps that deal with cleaning, and visualisation from a single instrument [@Salmon:2017]. In pleasant contrast, `mmaqshiny` integrates data from several devices along with location data, and offers unique correction algorithms for different instruments and parameters as described above.
 
