### 2020-06-28 #2250 @adithirgis
Hello @arfon ,

Both reviewers have accepted the package and paper in the current form and the package is now available on CRAN. Can we update the installation instructions in the paper.md accordingly at this stage?

Thanks!
