### 2020-06-28 #2250 @arfon
@adithirgis - could you please merge this PR which fixes some formatting issues with your paper: https://github.com/meenakshi-kushwaha/mmaqshiny/pull/4
