### 2020-06-28 #2250 @adithirgis
Hello @arfon!

I merged this PR - https://github.com/meenakshi-kushwaha/mmaqshiny/pull/4
Here is the Zenodo doi: 10.5281/zenodo.3911659

Thanks!
